Changed tls_version in tls_record to a scoped C++11 enum for better type-safety. Added funciton templates extract_most_significant_byte and extract_least_significant_byte to utils.hpp to  help operations on tls_version enum.

+++ src/io.cpp
@@ -336,7 +336,7 @@ namespace ntk {
         };
 
         print_field( "Content Type:", tls_content_type_names.at( record.content_type ) );
-        print_field( "Version:", record.version );
+        print_field( "Version:", static_cast<uint16_t>( record.version ) );
 
         std::cout << "===== TLS RECORD END =====\n\n";
     }
